amd_blobs: Add AMD's redistribution license in PDF and txt
We've created a new blobs repo just for AMD binaries that will all
be covered by this license.

This license was agreed to by AMD's lawyers for coreboot distribution.

I was hesitant about the clause that says not to download the binaries
before agreeing to the license, but figure we can put something in place
in coreboot to clone the repo as a bare repo, then checkout the files
after the user has agreed to the license either with a environment
variable or a Kconfig option.

BUG=b:131095312

Signed-off-by: Martin Roth <martinroth@chromium.org>
Change-Id: Iff2ee5a18a00d19654fb24514c62954af64f94e4