amd_blobs: Duplicate stoneyridge items
Begin migration of AMD assets from blobs to amd_blobs.  Rather than
adding a fresh unzip of the PI firmwares and VBIOS images, copy the
items directly from 3rdparty/blobs in case a bisect ever becomes
necessary.  The PSP blobs are placed into a directory structure that
better matches the PI's.

According to the log in 3rdparty/blobs//stoneyridge, these files are
mostly from release 1.3.0.9 with the following differences:
 * StoneyGenericVbios.006
 * An modified PSP bootloader
 * SMU firmware 33.9.0

Prior AMD license files are not duplicated here, and the license in
the top directory of this repo is in effect.

This change only covers firmware images released as part of the PI
package.  The derivative binaryPI image from AGESA source is copied
in a subsequent commit.

Change-Id: I9f730ee1f52985cc87a455c670d8803d3b888157
Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>